backend/requirements.txt
# Requirements for Django backend
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
gunicorn==23.0.0
Pillow==12.0.0
requests==2.32.5
google-generativeai==0.8.3
iyzipay==1.0.45
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
oauthlib==3.3.1
requests-oauthlib==2.0.0
